Move the `doc-client` module under the `@bayou` namespace.
