+++ client/js/TopControl.js
@@ -4,7 +4,7 @@
 
 import { SplitKey } from '@bayou/api-common';
 import { TheModule as appCommon_TheModule } from '@bayou/app-common';
-import { EditorComplex } from 'doc-client';
+import { EditorComplex } from '@bayou/doc-client';
 import { Logger } from '@bayou/see-all';
 import { TFunction, TObject } from '@bayou/typecheck';
