+++ local-modules/@bayou/data-model-client/package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
 
   "dependencies": {
-    "@bayou/deps-react": "local"
+    "@bayou/deps-react": "local",
+    "@bayou/doc-client": "local"
   }
 }
