+++ local-modules/@bayou/doc-client/CaretOverlay.js
