+++ local-modules/@bayou/doc-client/DocSession.js
