+++ local-modules/@bayou/doc-client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "doc-client",
+  "name": "@bayou/doc-client",
   "version": "1.0.0",
 
   "dependencies": {
